pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "geosam"
version = "0.1.0"
description = "Segment Anything Model for seismic interpretation"
readme = "README.md"
license = { file = "LICENSE" }
authors = [{ name = "Sainatha Yatham", email = "your@email.com" }]
requires-python = ">=3.10"
keywords = ["seismic", "segmentation", "geoscience", "SAM", "machine-learning"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Recognition",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

# Core dependencies — always installed
dependencies = [
    "numpy>=1.24",
    "segyio>=1.9",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "tqdm>=4.65",
]

# Optional extras — installed only when requested
[project.optional-dependencies]
sam = [
    "torch>=2.0",
    "torchvision>=0.15",
    "segment-anything-2",
    "opencv-python>=4.8",
    "Pillow>=9.5",
]
som = [
    "minisom>=2.3",
]
ui = [
    "gradio>=4.0",
]
all = [
    "geosam[sam]",
    "geosam[som]",
    "geosam[ui]",
]
modal = [
    "modal>=0.64",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.1",
]

[project.urls]
Homepage = "https://github.com/sainathyai/GeoSam"
Documentation = "https://github.com/sainathyai/GeoSam/docs"
"Bug Tracker" = "https://github.com/sainathyai/GeoSam/issues"
 